rename components to fix git issues
