+++ client/src/js/components/Partialss/Banner.jsx
