--- client/src/js/components/home/index.jsx
@@ -1,20 +1,0 @@
-import React from 'react';
-import Header from '../Partials/Header';
-import Footer from '../Partials/Footer';
-import HandPickedRecipe from '../Partials/HandPickedRecipe';
-import Banner from '../Partials/Banner';
-
-class HomePage extends React.Component {
-  render() {
-    return (
-      <div>
-        <Header />
-        <Banner />
-        <HandPickedRecipe />
-        <Footer />
-      </div>
-    );
-  }
-}
-
-export default HomePage;
